Formatting

+++ packages/wasmvm/wasmsolo/soloagent.go
@@ -34,21 +34,21 @@ func (a *SoloAgent) ScAgentID() wasmtypes.ScAgentID {
 	return wasmtypes.NewScAgentID(a.ScAddress(), wasmtypes.ScHname(a.hname))
 }
 
-func (a *SoloAgent) Balance(color ...wasmtypes.ScColor) int64 {
+func (a *SoloAgent) Balance(color ...wasmtypes.ScColor) uint64 {
 	switch len(color) {
 	case 0:
-		return int64(a.Env.GetAddressBalance(a.address, colored.IOTA))
+		return a.Env.GetAddressBalance(a.address, colored.IOTA)
 	case 1:
 		col, err := colored.ColorFromBytes(color[0].Bytes())
 		require.NoError(a.Env.T, err)
-		return int64(a.Env.GetAddressBalance(a.address, col))
+		return a.Env.GetAddressBalance(a.address, col)
 	default:
 		require.Fail(a.Env.T, "too many color arguments")
 		return 0
 	}
 }
 
-func (a *SoloAgent) Mint(amount int64) (wasmtypes.ScColor, error) {
-	color, err := a.Env.MintTokens(a.Pair, uint64(amount))
+func (a *SoloAgent) Mint(amount uint64) (wasmtypes.ScColor, error) {
+	color, err := a.Env.MintTokens(a.Pair, amount)
 	return wasmtypes.ColorFromBytes(color.Bytes()), err
 }
